Move the WizardUtils class to util package.
